+++ coloraide_extras/__init__.py
@@ -1,28 +1,30 @@
 """ColorAide Extra."""
 from .__meta__ import __version_info__, __version__  # noqa: F401
-from coloraide import NaN
-from coloraide.interpolate import Piecewise, Lerp
-from coloraide import Color as Base
-from .spaces.hsi import HSI
-from .spaces.ipt import IPT
-from .spaces.igpgtg import IgPgTg
-from .spaces.cmy import CMY
-from .spaces.cmyk import CMYK
+from coloraide import stop, hint, NaN
+from coloraide import ColorAll as Base
+from coloraide.types import Plugin
 from .spaces.ucs import UCS
 from .spaces.uvw import UVW
-from .spaces.xyy import XyY
-from .spaces.hunter_lab import HunterLab
-from .spaces.prismatic import Prismatic
-from .spaces.rlab import RLAB
-from .spaces.orgb import ORGB
+from .spaces.cam16_ucs import CAM16UCS, CAM16SCD, CAM16LCD
+from .distance.delta_e_cam16 import DECAM16
+from .contrast.contrast_weber import ContrastWeber
+from .contrast.contrast_michelson import ContrastMichelson
+from typing import List, Type
 
-__all__ = ("Color", "SPACES", 'NaN', 'Piecewise', 'Lerp')
+__all__ = ("Color", "SPACES", 'NaN', 'stop', 'hint')
 
-SPACES = [HSI, IPT, IgPgTg, CMY, CMYK, UCS, UVW, XyY, HunterLab, Prismatic, RLAB, ORGB]
+SPACES = [UCS, UVW, CAM16UCS, CAM16SCD, CAM16LCD]  # type: List[Type[Plugin]]
+
+DE = [DECAM16]  # type: List[Type[Plugin]]
+
+CONTRAST = [ContrastWeber, ContrastMichelson]  # type: List[Type[Plugin]]
 
 
 class Color(Base):
     """Color class containing all default and extra color spaces."""
 
 
-Color.register(SPACES)
+# If we get out of sync with ColorAide, we may try
+# and overwrite a package that used to be here, but
+# is now hosted there.
+Color.register(SPACES + DE + CONTRAST, silent=True)