+++ coloraide_extras/distance/__init__.py
@@ -1,0 +1,1 @@
+"""Distancing algorithms."""