+++ coloraide_extras/spaces/ucs.py
@@ -4,7 +4,7 @@ CIE 1960 UCS color class.
 http://en.wikipedia.org/wiki/CIE_1960_color_space#Relation_to_CIE_XYZ
 """
 from coloraide.spaces import Space
-from coloraide.gamut.bounds import GamutUnbound
+from coloraide.channels import Channel
 from coloraide.cat import WHITES
 from coloraide.types import Vector
 from typing import Tuple
@@ -30,50 +30,12 @@ class UCS(Space):
     BASE = "xyz-d65"
     NAME = "ucs"
     SERIALIZE = ("--ucs",)  # type: Tuple[str, ...]
-    CHANNEL_NAMES = ("u", "v", "w")
-    WHITE = WHITES['2deg']['D65']
-
-    BOUNDS = (
-        GamutUnbound(0.0, 1.0),
-        GamutUnbound(0.0, 1.0),
-        GamutUnbound(0.0, 1.0)
+    CHANNELS = (
+        Channel("u", 0.0, 1.0),
+        Channel("v", 0.0, 1.0),
+        Channel("w", 0.0, 1.0)
     )
-
-    @property
-    def u(self) -> float:
-        """The u channel."""
-
-        return self._coords[0]
-
-    @u.setter
-    def u(self, value: float) -> None:
-        """Set u."""
-
-        self._coords[0] = value
-
-    @property
-    def v(self) -> float:
-        """The v channel."""
-
-        return self._coords[1]
-
-    @v.setter
-    def v(self, value: float) -> None:
-        """Set v."""
-
-        self._coords[1] = value
-
-    @property
-    def w(self) -> float:
-        """The w channel."""
-
-        return self._coords[2]
-
-    @w.setter
-    def w(self, value: float) -> None:
-        """Set w."""
-
-        self._coords[2] = value
+    WHITE = WHITES['2deg']['D65']
 
     @classmethod
     def to_base(cls, coords: Vector) -> Vector: