+++ tools/buildwheel.py
@@ -12,13 +12,13 @@ import hashlib
 # Notebook specific wheels
 NOTEBOOK_WHEELS = [
     "https://files.pythonhosted.org/packages/6e/33/1ae0f71395e618d6140fbbc9587cc3156591f748226075e0f7d6f9176522/Markdown-3.3.4-py3-none-any.whl",  # noqa: E501
-    "https://files.pythonhosted.org/packages/37/28/8750666d9431c9cb5201df4bfa14b5ffcdcf659b22156a4810ee9ad5ca06/pymdown_extensions-9.4-py3-none-any.whl",  # noqa: E501
+    "https://files.pythonhosted.org/packages/f1/e0/1ed09f66cd1648f8e009120debf9b7d67596fb688e53e71522da1daa02a0/pymdown_extensions-9.5-py3-none-any.whl",  # noqa: E501
 ]
 
 # Wheels required in addition to the current project
 PLAYGROUND_WHEELS = [
     "https://files.pythonhosted.org/packages/5c/8e/1d9017950034297fffa336c72e693a5b51bbf85141b24a763882cf1977b5/Pygments-2.12.0-py3-none-any.whl",  # noqa: E501
-    "https://files.pythonhosted.org/packages/6d/61/e2f04ad470b7eb312e3ad2d3dcc70150be2dc6d4789d0ee3c0bf5018e56e/coloraide-0.16.0-py3-none-any.whl"  # noqa: E501
+    "https://files.pythonhosted.org/packages/cc/0f/db76941e707138fe60728a6e7a322da9f62b2ed9a70cb2b9eb3f1641eeb8/coloraide-1.0b3-py3-none-any.whl"  # noqa: E501
 ]
 
 MKDOCS_YML = 'mkdocs.yml'
@@ -30,7 +30,7 @@ CONFIG = """\
 var color_notebook = {{
     "playground_wheels": {},
     "notebook_wheels": {},
-    "default_playground": "from coloraide_extras import Color\\ncoloraide.__version__\\nColor('color(--xyy 0.64 0.33 0.21264)')"
+    "default_playground": "from coloraide_extras import Color\\ncoloraide.__version__\\nColor('color(--ucs 0.27493 0.21264 0.12243 / 1)')"
 }}
 """  # noqa: E501